requirements.txt
# Core ML/DL
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Reinforcement Learning
gymnasium>=0.28.0

# Feature Engineering
# Note: ta-lib requires system library installation
# On macOS: brew install ta-lib
# ta-lib>=0.4.25

# Alternative technical indicators (no C dependencies)
pandas-ta>=0.3.14b

# Gradient Boosting (for feature selection)
xgboost>=1.7.0

# Data Storage
duckdb>=0.8.0

# Async Networking
websockets>=11.0
aiohttp>=3.8.0

# Bybit API
pybit>=5.6.0

# Reddit API
praw>=7.7.0

# Visualization & Logging
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.5.0

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
